Change tabs to Spaces (#27)
This is a revolutionary change, one can argue this is the best diff since the inception of this project.
Tabs are ugly, I use to like them but no more, fixing a tab from 4 to 2 in your editor still doesn't make any sense, sure it saves extra bits
But this is what anarchy looks like. So.